tests/submission/Microsoft/semantic/aside.html

<!doctype html>
<html>
<head>
 <meta content="text/html; charset=utf-8" http-equiv="content-type" />
 <title>verify ASIDE is a HTMLElement</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
 <aside id="aside">
 This is an ASIDE with a NAV in it
 <nav>
 <ul>
  <li><a href="#">some test link</a>
  <li><a href="#">another test link</a>
 </ul>
 </nav>
</aside>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("aside").toString() )
 })
</script>
</body>
</html>
